audio: Fix issues during device switch
Issue 1: On L, checkDeviceMuteStrategies is applying
temporary mute on current output only. It does not
mute other active outputs. This is leading to glitches
when touch tones played with music

Issue 2: checkDeviceMuteStrategies considers latency
of all active outputs to compute mute delay for the
current output.If direct output on A2DP device is
active then start up latency is more for rringtones

Fix is to apply temporary mute for other active outputs
and calculate mute delay based on the latency of the
current output

CRs-Fixed: 774775

Change-Id: Ic0880df5fee631f588578137618a8fa51c03ad6d